+++ src/components/ExitModal/ExitModal.js
@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import React from 'react';
 import { createPortal } from 'react-dom';
 import css from './exitModal.module.css';
