+++ src/pages/auth/RegistrationForm.module.css
@@ -114,4 +114,5 @@
 
 .link {
   color: var(--main-text-color);
+  text-decoration: none;
 }
