+++ src/redux/store.js
@@ -18,7 +18,7 @@ import storage from 'redux-persist/lib/storage';
 import { authSlice } from './auth/authReducer';
 
 const persistConfig = {
-  key: 'trecker',
+  key: 'root',
   storage,
   // whitelist: ['']
 };
@@ -34,13 +34,13 @@ const middleware = [
 
 const rootReducer = combineReducers({
   [authSlice.name]: authSlice.reducer,
-  middleware,
 });
 
 const persistedReducer = persistReducer(persistConfig, rootReducer);
 
 export const store = configureStore({
   reducer: persistedReducer,
+  middleware,
 });
 
 export const persistor = persistStore(store);
